"Models" module added
When I started this, I downloaded from master, days before 2.6 release.

I was creating the bindings using the cheatsheet from 2.5, however, the
cheatsheet suddenly updated to 3.0 and I started to bind to 3.0
accidentally.

So, is possible that my wrapper is actually a mess, however, the
examples runs :v

I'll follow the 3.0 cheatsheet, and put a special focus on examples to
test the wrapper, then I'll remake the previous modules.