WIP raylib header reader
The goal is read the raylib.h (and raymath.h in the future) to generate
a table with all definitions, for example:
{ -- [...]
   args = { {
      name = "size",
      type = { "int" }
   } },
   comment = "Default size for new audio streams",
   name = "SetAudioStreamBufferSizeDefault",
   return_type = { "void" }
}

how to use (while it's in WIP):
$ cd binding_generator
$ lua header_reader.lua

for now, it prints a lot of things, the last one is the generated
structs and RLAPIs tables.

when is done, a binding_generator.lua script will load and run
header_header.lua and then it will generate the raylib.nelua file.

Well, it's possible to use nelua's metaprogramming, however I think
generating a raylib.nelua file is a better option because it works as
the raylib's cheatsheet
(https://www.raylib.com/cheatsheet/cheatsheet.html).